+++ db/upgrade.php
@@ -138,5 +138,25 @@ function xmldb_playerwords_upgrade(int $oldversion): bool {
         upgrade_mod_savepoint(true, 2026050805, 'playerwords');
     }
 
+    if ($oldversion < 2026050811) {
+        $table = new xmldb_table('playerwords');
+
+        $glossaryid = new xmldb_field(
+            'glossaryid',
+            XMLDB_TYPE_INTEGER,
+            '10',
+            null,
+            XMLDB_NOTNULL,
+            null,
+            '0',
+            'aigranularity'
+        );
+        if (!$dbman->field_exists($table, $glossaryid)) {
+            $dbman->add_field($table, $glossaryid);
+        }
+
+        upgrade_mod_savepoint(true, 2026050811, 'playerwords');
+    }
+
     return true;
 }